requirements.txt
beautifulsoup4==4.8.2
lxml==4.3.3
PyYAML==5.3.1
mitmproxy==4.0.4
py3shark==0.4.0
requests==2.21.0
requests-toolbelt==0.9.1
redis==3.4.1
